apply MVC pattern to keymap editor code
