+++ PlayTools/Editor/Controllers/Elements/Instances/ParentElement.swift
@@ -1,0 +1,12 @@
+//
+//  ParentElement.swift
+//  PlayTools
+//
+//  Created by 许沂聪 on 2024/6/2.
+//
+
+import Foundation
+
+protocol ParentElement: ControlModel {
+    func unfocusChildren()
+}
