+++ PlayTools/Keymap/Keymapping.swift
@@ -63,47 +63,3 @@ class Keymapping {
         }
     }
 }
-// Data structure definition should match those in
-// https://github.com/PlayCover/PlayCover/blob/develop/PlayCover/Model/Keymapping.swift
-struct KeyModelTransform: Codable {
-    var size: CGFloat
-    var xCoord: CGFloat
-    var yCoord: CGFloat
-}
-// controller buttons are indexed with names
-struct Button: Codable {
-    var keyCode: Int
-    var keyName: String
-    var transform: KeyModelTransform
-}
-
-struct Joystick: Codable {
-    var upKeyCode: Int
-    var rightKeyCode: Int
-    var downKeyCode: Int
-    var leftKeyCode: Int
-    var keyName: String = "Keyboard"
-    var transform: KeyModelTransform
-}
-
-struct MouseArea: Codable {
-    var keyName: String
-    var transform: KeyModelTransform
-    init(transform: KeyModelTransform) {
-        self.transform = transform
-        self.keyName = "Mouse"
-    }
-    init(keyName: String, transform: KeyModelTransform) {
-        self.transform = transform
-        self.keyName = keyName
-    }
-}
-
-struct KeymappingData: Codable {
-    var buttonModels: [Button] = []
-    var draggableButtonModels: [Button] = []
-    var joystickModel: [Joystick] = []
-    var mouseAreaModel: [MouseArea] = []
-    var bundleIdentifier: String
-    var version = "2.0.0"
-}
